move the arcgis link to its own component
