+++ app/components/geoblacklight/arc_gis_component.html.erb
@@ -1,4 +1,3 @@
-<% document ||= @document %>
 <%= link_to(arcgis_link(document.arcgis_urls)) do %>
   <%= blacklight_icon('esri-globe') %><%= t('geoblacklight.tools.open_arcgis') %>
 <% end %>
